+++ src/test/java/com/lazerycode/jmeter/testrunner/ResultScannerTest.java
@@ -5,6 +5,8 @@ import org.junit.Test;
 
 import java.io.File;
 import java.net.URL;
+import java.util.ArrayList;
+import java.util.List;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
@@ -14,6 +16,9 @@ public class ResultScannerTest {
     private static final boolean DO_NOT_COUNT_FAILURES = false;
     private static final boolean COUNT_SUCCESSES = true;
     private static final boolean DO_NOT_COUNT_SUCCESSES = false;
+    private static final boolean DEFAULT_IS_CSV = false;
+    private static final boolean DEFAULT_ONLY_FAIL_WHEN_MATCHING_FAILURE_MESSAGES = false;
+    private static final List<String> DEFAULT_FAILURE_LIST = new ArrayList<>();
     private final URL jtlFailingResultsFileURL = this.getClass().getResource("/jtl2-1-fail.jtl");
     private final URL jtlPassingResultsFileURL = this.getClass().getResource("/jtl2-1-pass.jtl");
     private final URL csvFailingResultsFileURL = this.getClass().getResource("/csv2-1-fail.csv");
@@ -25,7 +30,7 @@ public class ResultScannerTest {
     @Test
     public void jtlFileWithFailuresCountSuccessAndFailures() throws Exception {
         File resultsFile = new File(jtlFailingResultsFileURL.toURI());
-        ResultScanner fileScanner = new ResultScanner(COUNT_SUCCESSES, COUNT_FAILURES);
+        ResultScanner fileScanner = new ResultScanner(COUNT_SUCCESSES, COUNT_FAILURES, DEFAULT_IS_CSV, DEFAULT_ONLY_FAIL_WHEN_MATCHING_FAILURE_MESSAGES, DEFAULT_FAILURE_LIST);
         fileScanner.parseResultFile(resultsFile);
 
         assertThat(fileScanner.getFailureCount()).isEqualTo(2);
@@ -35,7 +40,7 @@ public class ResultScannerTest {
     @Test
     public void jtlFileWithFailuresCountSuccessesOnly() throws Exception {
         File resultsFile = new File(jtlFailingResultsFileURL.toURI());
-        ResultScanner fileScanner = new ResultScanner(COUNT_SUCCESSES, DO_NOT_COUNT_FAILURES);
+        ResultScanner fileScanner = new ResultScanner(COUNT_SUCCESSES, DO_NOT_COUNT_FAILURES, DEFAULT_IS_CSV, DEFAULT_ONLY_FAIL_WHEN_MATCHING_FAILURE_MESSAGES, DEFAULT_FAILURE_LIST);
         fileScanner.parseResultFile(resultsFile);
 
         assertThat(fileScanner.getFailureCount()).isEqualTo(0);
@@ -45,7 +50,7 @@ public class ResultScannerTest {
     @Test
     public void jtlFileWithFailuresCountFailuresOnly() throws Exception {
         File resultsFile = new File(jtlFailingResultsFileURL.toURI());
-        ResultScanner fileScanner = new ResultScanner(DO_NOT_COUNT_SUCCESSES, COUNT_FAILURES);
+        ResultScanner fileScanner = new ResultScanner(DO_NOT_COUNT_SUCCESSES, COUNT_FAILURES, DEFAULT_IS_CSV, DEFAULT_ONLY_FAIL_WHEN_MATCHING_FAILURE_MESSAGES, DEFAULT_FAILURE_LIST);
         fileScanner.parseResultFile(resultsFile);
 
         assertThat(fileScanner.getFailureCount()).isEqualTo(2);
@@ -55,7 +60,7 @@ public class ResultScannerTest {
     @Test
     public void jtlFileWithNoFailuresCountSuccessAndFailures() throws Exception {
         File resultsFile = new File(jtlPassingResultsFileURL.toURI());
-        ResultScanner fileScanner = new ResultScanner(COUNT_SUCCESSES, COUNT_FAILURES);
+        ResultScanner fileScanner = new ResultScanner(COUNT_SUCCESSES, COUNT_FAILURES, DEFAULT_IS_CSV, DEFAULT_ONLY_FAIL_WHEN_MATCHING_FAILURE_MESSAGES, DEFAULT_FAILURE_LIST);
         fileScanner.parseResultFile(resultsFile);
 
         assertThat(fileScanner.getFailureCount()).isEqualTo(0);
@@ -65,7 +70,7 @@ public class ResultScannerTest {
     @Test
     public void jtlFileWithNoFailuresCountSuccessesOnly() throws Exception {
         File resultsFile = new File(jtlPassingResultsFileURL.toURI());
-        ResultScanner fileScanner = new ResultScanner(COUNT_SUCCESSES, DO_NOT_COUNT_FAILURES);
+        ResultScanner fileScanner = new ResultScanner(COUNT_SUCCESSES, DO_NOT_COUNT_FAILURES, DEFAULT_IS_CSV, DEFAULT_ONLY_FAIL_WHEN_MATCHING_FAILURE_MESSAGES, DEFAULT_FAILURE_LIST);
         fileScanner.parseResultFile(resultsFile);
 
         assertThat(fileScanner.getFailureCount()).isEqualTo(0);
@@ -75,7 +80,7 @@ public class ResultScannerTest {
     @Test
     public void jtlFileWithNoFailuresCountFailuresOnly() throws Exception {
         File resultsFile = new File(jtlPassingResultsFileURL.toURI());
-        ResultScanner fileScanner = new ResultScanner(DO_NOT_COUNT_SUCCESSES, COUNT_FAILURES);
+        ResultScanner fileScanner = new ResultScanner(DO_NOT_COUNT_SUCCESSES, COUNT_FAILURES, DEFAULT_IS_CSV, DEFAULT_ONLY_FAIL_WHEN_MATCHING_FAILURE_MESSAGES, DEFAULT_FAILURE_LIST);
         fileScanner.parseResultFile(resultsFile);
 
         assertThat(fileScanner.getFailureCount()).isEqualTo(0);
@@ -85,7 +90,7 @@ public class ResultScannerTest {
     @Test
     public void csvFileWithFailuresCountSuccessAndFailures() throws Exception {
         File resultsFile = new File(csvFailingResultsFileURL.toURI());
-        ResultScanner fileScanner = new ResultScanner(COUNT_SUCCESSES, COUNT_FAILURES, true);
+        ResultScanner fileScanner = new ResultScanner(COUNT_SUCCESSES, COUNT_FAILURES, true, DEFAULT_ONLY_FAIL_WHEN_MATCHING_FAILURE_MESSAGES, DEFAULT_FAILURE_LIST);
         fileScanner.parseResultFile(resultsFile);
 
         assertThat(fileScanner.getFailureCount()).isEqualTo(2);
@@ -95,7 +100,7 @@ public class ResultScannerTest {
     @Test
     public void csvFileWithNoFailuresCountSuccessAndFailures() throws Exception {
         File resultsFile = new File(csvPassingResultsFileURL.toURI());
-        ResultScanner fileScanner = new ResultScanner(COUNT_SUCCESSES, COUNT_FAILURES, true);
+        ResultScanner fileScanner = new ResultScanner(COUNT_SUCCESSES, COUNT_FAILURES, true, DEFAULT_ONLY_FAIL_WHEN_MATCHING_FAILURE_MESSAGES, DEFAULT_FAILURE_LIST);
         fileScanner.parseResultFile(resultsFile);
 
         assertThat(fileScanner.getFailureCount()).isEqualTo(0);
@@ -105,7 +110,7 @@ public class ResultScannerTest {
     @Test
     public void csvFileWithNoFailuresCountSuccessAndFailuresAlternateSep() throws Exception {
         File resultsFile = new File(csvWithAlternateSeparatorPassingResultsFileURL.toURI());
-        ResultScanner fileScanner = new ResultScanner(COUNT_SUCCESSES, COUNT_FAILURES, true);
+        ResultScanner fileScanner = new ResultScanner(COUNT_SUCCESSES, COUNT_FAILURES, true, DEFAULT_ONLY_FAIL_WHEN_MATCHING_FAILURE_MESSAGES, DEFAULT_FAILURE_LIST);
         fileScanner.parseResultFile(resultsFile);
 
         assertThat(fileScanner.getFailureCount()).isEqualTo(0);
@@ -115,35 +120,72 @@ public class ResultScannerTest {
     @Test(expected = IllegalArgumentException.class)
     public void emptyCSVFileThrowsIllegalArgumentException() throws Exception {
         File resultsFile = new File(emptyCSVFileURL.toURI());
-        ResultScanner fileScanner = new ResultScanner(COUNT_SUCCESSES, COUNT_FAILURES, true);
+        ResultScanner fileScanner = new ResultScanner(COUNT_SUCCESSES, COUNT_FAILURES, true, DEFAULT_ONLY_FAIL_WHEN_MATCHING_FAILURE_MESSAGES, DEFAULT_FAILURE_LIST);
         fileScanner.parseResultFile(resultsFile);
     }
 
     @Test(expected = IllegalStateException.class)
     public void csvFileMissingDelimiterThrowsIllegalStateException() throws Exception {
         File resultsFile = new File(csvMissingDelimiterFileURL.toURI());
-        ResultScanner fileScanner = new ResultScanner(COUNT_SUCCESSES, COUNT_FAILURES, true);
+        ResultScanner fileScanner = new ResultScanner(COUNT_SUCCESSES, COUNT_FAILURES, true, DEFAULT_ONLY_FAIL_WHEN_MATCHING_FAILURE_MESSAGES, DEFAULT_FAILURE_LIST);
         fileScanner.parseResultFile(resultsFile);
+    }
+
+    @Test
+    public void csvFileOnlyCountMatchingFailureMessages() throws Exception {
+        File resultsFile = new File(csvFailingResultsFileURL.toURI());
+        List<String> failureMessages = new ArrayList<>();
+        failureMessages.add("It went wrong!");
+        ResultScanner fileScanner = new ResultScanner(COUNT_SUCCESSES, COUNT_FAILURES, true, true, failureMessages);
+        fileScanner.parseResultFile(resultsFile);
+
+        assertThat(fileScanner.getFailureCount()).isEqualTo(1);
+        assertThat(fileScanner.getSuccessCount()).isEqualTo(0);
+    }
+
+    @Test
+    public void csvFileOnlyCountMatchingFailureMessagesThatAreNotFound() throws Exception {
+        File resultsFile = new File(csvFailingResultsFileURL.toURI());
+        List<String> failureMessages = new ArrayList<>();
+        failureMessages.add("FailureMessageNotFound");
+        ResultScanner fileScanner = new ResultScanner(COUNT_SUCCESSES, COUNT_FAILURES, true, true, failureMessages);
+        fileScanner.parseResultFile(resultsFile);
+
+        assertThat(fileScanner.getFailureCount()).isEqualTo(0);
+        assertThat(fileScanner.getSuccessCount()).isEqualTo(0);
+    }
+
+    @Test
+    public void csvFileOnlyCountMatchingFailureMessagesWhenSomeAreFound() throws Exception {
+        File resultsFile = new File(csvFailingResultsFileURL.toURI());
+        List<String> failureMessages = new ArrayList<>();
+        failureMessages.add("It went wrong!");
+        failureMessages.add("FailureMessageNotFound");
+        ResultScanner fileScanner = new ResultScanner(COUNT_SUCCESSES, COUNT_FAILURES, true, true, failureMessages);
+        fileScanner.parseResultFile(resultsFile);
+
+        assertThat(fileScanner.getFailureCount()).isEqualTo(1);
+        assertThat(fileScanner.getSuccessCount()).isEqualTo(0);
     }
 
     @Test(expected = MojoExecutionException.class)
     public void fileThatDoesNotExistThrowsResultsFileNotFoundException() throws Exception {
         File resultsFile = new File("DoesNotExist.nope");
-        ResultScanner fileScanner = new ResultScanner(COUNT_SUCCESSES, COUNT_FAILURES, true);
+        ResultScanner fileScanner = new ResultScanner(COUNT_SUCCESSES, COUNT_FAILURES, true, DEFAULT_ONLY_FAIL_WHEN_MATCHING_FAILURE_MESSAGES, DEFAULT_FAILURE_LIST);
         fileScanner.parseResultFile(resultsFile);
     }
 
     @Test(expected = MojoExecutionException.class)
     public void invalidCSVFileThrowsIOException() throws Exception {
         File resultsFile = new File("/");
-        ResultScanner fileScanner = new ResultScanner(COUNT_SUCCESSES, COUNT_FAILURES, true);
+        ResultScanner fileScanner = new ResultScanner(COUNT_SUCCESSES, COUNT_FAILURES, true, DEFAULT_ONLY_FAIL_WHEN_MATCHING_FAILURE_MESSAGES, DEFAULT_FAILURE_LIST);
         fileScanner.parseResultFile(resultsFile);
     }
 
     @Test(expected = MojoExecutionException.class)
     public void invalidJTLileThrowsIOException() throws Exception {
         File resultsFile = new File("/");
-        ResultScanner fileScanner = new ResultScanner(COUNT_SUCCESSES, COUNT_FAILURES, false);
+        ResultScanner fileScanner = new ResultScanner(COUNT_SUCCESSES, COUNT_FAILURES, false, DEFAULT_ONLY_FAIL_WHEN_MATCHING_FAILURE_MESSAGES, DEFAULT_FAILURE_LIST);
         fileScanner.parseResultFile(resultsFile);
     }